# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fellowship Entrance Exam and Seat Allocation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates/{candidateId}/exam-assignments:
    get:
      operationId: listCandidateExamAssignments
      tags: [candidates]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: candidateId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CandidateExamAssignment" }
    put:
      operationId: setCandidateExamAssignments
      tags: [candidates]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: candidateId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetExamAssignmentsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CandidateExamAssignment" }
...
  /documents/me:
    get:
      operationId: listMyDocuments
      tags: [documents]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
    post:
      operationId: uploadMyDocument
      tags: [documents]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: programId
          in: query
          schema: { type: integer }
        - name: kind
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Exam" }
    post:
      operationId: createExam
      tags: [exams]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewExam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Exam" }
  /exams/{examId}:
    get:
      operationId: getExam
      tags: [exams]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: examId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExamDetail" }
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Doctor" }
  /doctors/{doctorId}/assign-unit:
    post:
      operationId: assignDoctorUnit
      tags: [doctors]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: doctorId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorUnitAssignment"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
  /doctors/{doctorId}/assign-candidate:
    post:
      operationId: assignDoctorCandidate
      tags: [doctors]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: doctorId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorCandidateAssignment"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InterviewAssignment" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CandidateExamAssignment:
      type: object
      properties:
        id: { type: integer }
        examId: { type: integer }
        examTitle: { type: string }
        examKind: { type: string }
        assignedAt: { type: string }
      required: [id, examId, examTitle, examKind, assignedAt]
    SetExamAssignmentsBody:
      type: object
      properties:
        examIds:
          type: array
          items: { type: integer }
      required: [examIds]
...
    Document:
      type: object
      properties:
        id: { type: integer }
        docType: { type: string }
        fileName: { type: string }
        uploadedAt: { type: string }
      required: [id, docType, fileName, uploadedAt]
    UploadDocumentBody:
      type: object
      properties:
        docType: { type: string }
        fileName: { type: string }
        contentBase64: { type: string }
      required: [docType, fileName, contentBase64]
    Exam:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        kind: { type: string }
        programId: { type: integer, nullable: true }
        programName: { type: string, nullable: true }
        durationMinutes: { type: integer }
        totalQuestions: { type: integer }
        questionCount: { type: integer }
        passingScore: { type: number, nullable: true }
        active: { type: boolean }
        startsAt: { type: string, nullable: true }
        endsAt: { type: string, nullable: true }
      required: [id, title, kind, durationMinutes, totalQuestions, questionCount, active]
    ExamDetail:
      allOf:
        - $ref: "#/components/schemas/Exam"
        - type: object
          properties:
            description: { type: string, nullable: true }
    NewExam:
      type: object
      properties:
        title: { type: string }
        kind: { type: string }
        programId: { type: integer, nullable: true }
        durationMinutes: { type: integer }
        totalQuestions: { type: integer }
        passingScore: { type: number }
        description: { type: string }
        startsAt: { type: string }
        endsAt: { type: string }
      required: [title, kind, durationMinutes, totalQuestions]
...
    InterviewAssignment:
      type: object
      properties:
        id: { type: integer }
        candidateId: { type: integer }
        candidateName: { type: string }
        candidateCode: { type: string }
        unitName: { type: string, nullable: true }
        scheduledAt: { type: string, nullable: true }
        status: { type: string }
        existingScore:
          $ref: "#/components/schemas/InterviewScore"
          nullable: true
      required: [id, candidateId, candidateName, candidateCode, status]
    InterviewScore:
      type: object
      properties:
        id: { type: integer }
        candidateId: { type: integer }
        doctorId: { type: integer }
        doctorName: { type: string }
        score: { type: number }
        remarks: { type: string, nullable: true }
        submittedAt: { type: string }
      required: [id, candidateId, doctorId, doctorName, score, submittedAt]
...
    Doctor:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        fullName: { type: string }
        email: { type: string }
        unitId: { type: integer, nullable: true }
        unitName: { type: string, nullable: true }
        assignedCandidates: { type: integer }
      required: [id, userId, fullName, email, assignedCandidates]
    DoctorUnitAssignment:
      type: object
      properties:
        unitId: { type: integer }
      required: [unitId]
    DoctorCandidateAssignment:
      type: object
      properties:
        candidateId: { type: integer }
      required: [candidateId]
...